fix: remove utils and simplify crates
