+++ examples/actix-web/src/main.rs
@@ -3,7 +3,7 @@ use actix_web::{web, App, HttpResponse, HttpServer};
 use anyhow::Result;
 use apalis::prelude::*;
 use apalis::utils::TokioExecutor;
-use apalis::{layers::TraceLayer, redis::RedisStorage};
+use apalis::{layers::tracing::TraceLayer, redis::RedisStorage};
 use futures::future;
 
 use email_service::{send_email, Email};
@@ -26,7 +26,8 @@ async fn main() -> Result<()> {
     std::env::set_var("RUST_LOG", "debug");
     env_logger::init();
 
-    let storage = RedisStorage::connect("redis://127.0.0.1/").await?;
+    let conn = apalis::redis::connect("redis://127.0.0.1/").await?;
+    let storage = RedisStorage::new(conn);
     let data = web::Data::new(storage.clone());
     let http = async {
         HttpServer::new(move || {
