+++ examples/async-std-runtime/Cargo.toml
@@ -7,9 +7,14 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1"
-apalis = { path = "../../", default-features = false, features = ["cron", "async-std-comp", "tracing", "retry"] }
-apalis-utils = { path = "../../packages/apalis-utils", default-features = false, features = ["sleep"] }
-async-std = { version = "1.12.0", features=["attributes"] }
+apalis = { path = "../../", default-features = false, features = [
+    "cron",
+    "async-std-comp",
+    "tracing",
+    "retry",
+] }
+apalis-core = { path = "../../packages/apalis-core", default-features = false }
+async-std = { version = "1.12.0", features = ["attributes"] }
 serde = "1"
 tracing-subscriber = "0.3.11"
 chrono = { version = "0.4", default-features = false, features = ["clock"] }
