+++ examples/mysql/src/main.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use apalis::mysql::MySqlPool;
 use apalis::prelude::*;
-use apalis::{layers::TraceLayer, mysql::MysqlStorage};
+use apalis::{layers::tracing::TraceLayer, mysql::MysqlStorage};
 use email_service::{send_email, Email};
 
 async fn produce_jobs(storage: &MysqlStorage<Email>) -> Result<()> {
