+++ examples/prometheus/src/main.rs
@@ -5,7 +5,7 @@
 //! ```
 use anyhow::Result;
 use apalis::prelude::*;
-use apalis::{layers::PrometheusLayer, redis::RedisStorage};
+use apalis::{layers::prometheus::PrometheusLayer, redis::RedisStorage};
 use axum::{
     extract::Form,
     http::StatusCode,
@@ -29,7 +29,8 @@ async fn main() -> Result<()> {
         ))
         .with(tracing_subscriber::fmt::layer())
         .init();
-    let storage: RedisStorage<Email> = RedisStorage::connect("redis://127.0.0.1/").await?;
+    let conn = apalis::redis::connect("redis://127.0.0.1/").await?;
+    let storage = RedisStorage::new(conn);
     // build our application with some routes
     let recorder_handle = setup_metrics_recorder();
     let app = Router::new()
