+++ examples/redis/Cargo.toml
@@ -8,9 +8,7 @@ license = "MIT OR Apache-2.0"
 [dependencies]
 anyhow = "1"
 tokio = { version = "1", features = ["full"] }
-apalis = { path = "../../" }
-apalis-redis = { path = "../../packages/apalis-redis" }
-apalis-utils = { path = "../../packages/apalis-utils" }
+apalis = { path = "../../", features = ["redis"]}
 serde = "1"
 env_logger = "0.10"
 tracing-subscriber = "0.3.11"
