+++ examples/sqlite/Cargo.toml
@@ -8,8 +8,10 @@ license = "MIT OR Apache-2.0"
 [dependencies]
 anyhow = "1"
 tokio = { version = "1", features = ["full"] }
-apalis = { path = "../../", features = ["sqlite", "limit", "tracing"] }
-apalis-utils = { path = "../../packages/apalis-utils", features = [
+apalis = { path = "../../", features = [
+    "sqlite",
+    "limit",
+    "tracing",
     "tokio-comp",
 ] }
 serde = { version = "1", features = ["derive"] }
