+++ packages/apalis-redis/Cargo.toml
@@ -11,7 +11,10 @@ description = "Redis Storage for apalis: use Redis for background jobs and messa
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-apalis-core = { path = "../../packages/apalis-core", version = "0.5.0-alpha.0", default-features = false }
+apalis-core = { path = "../../packages/apalis-core", version = "0.5.0-alpha.0", default-features = false, features = [
+    "sleep",
+    "json",
+] }
 redis = { version = "0.24", default-features = false, features = [
     "script",
     "aio",
@@ -24,15 +27,12 @@ chrono = { version = "0.4", default-features = false, features = [
     "serde",
 ] }
 time = { version = "0.3", optional = true, features = ["serde"] }
-serde_json = "1"
+
 async-stream = "0.3"
 futures = "0.3"
 tokio = { version = "1", features = ["rt", "net"], optional = true }
 async-std = { version = "1.12.0", optional = true }
 async-trait = "0.1.77"
-apalis-utils = { path = "../apalis-utils", version = "0.1.0", features = [
-    "sleep",
-] }
 
 
 [dev-dependencies]
