+++ packages/apalis-sql/Cargo.toml
@@ -25,11 +25,11 @@ features = ["chrono"]
 [dependencies]
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-apalis-core = { path = "../../packages/apalis-core", version = "0.5.0-alpha.0", default-features = false }
-apalis-utils = { path = "../../packages/apalis-utils", version = "0.1.0", default-features = false, features = [
+apalis-core = { path = "../../packages/apalis-core", version = "0.5.0-alpha.0", default-features = false, features = [
     "sleep",
     "json",
 ] }
+
 futures = "0.3"
 async-stream = "0.3"
 tokio = { version = "1", features = ["rt", "net"], optional = true }
