+++ packages/apalis-sql/src/context.rs
@@ -1,6 +1,6 @@
 use apalis_core::error::Error;
+use apalis_core::task::{attempt::Attempt, task_id::TaskId};
 use apalis_core::worker::WorkerId;
-use apalis_utils::{attempt::Attempt, task_id::TaskId};
 use serde::{Deserialize, Serialize};
 use sqlx::types::chrono::{DateTime, Utc};
 use std::{fmt, str::FromStr, time::SystemTime};
