--- packages/apalis-utils/Cargo.toml
@@ -1,53 +1,0 @@
-[package]
-name = "apalis-utils"
-version = "0.1.0"
-edition = "2021"
-
-[features]
-default = ["tokio-comp", "layers", "json"]
-sleep = ["async-timer"]
-tokio-comp = ["tokio"]
-async-std-comp = ["async-std"]
-layers = [
-    "sentry",
-    "prometheus",
-    "trace",
-    "retry",
-    "timeout",
-    "limit",
-    "filter",
-]
-sentry = ["sentry-core", "ulid/uuid", "uuid"]
-prometheus = ["metrics", "metrics-exporter-prometheus"]
-trace = ["tracing-futures/std-future"]
-retry = ["tower/retry"]
-timeout = ["tower/timeout"]
-limit = ["tower/limit"]
-filter = ["tower/filter"]
-
-json = ["serde_json"]
-
-[dependencies]
-async-timer = { version = "1.0.0-beta.11", optional = true, default-features = false }
-apalis-core = { version = "0.5.0-alpha.0", path = "../apalis-core" }
-tokio = { version = "1", features = [
-    "rt",
-], default-features = false, optional = true }
-async-std = { version = "1", optional = true }
-tower = { version = "0.4", features = ["util"], default-features = false }
-tracing-futures = { version = "0.2.5", optional = true, default-features = false }
-sentry-core = { version = "0.32.1", optional = true, default-features = false }
-metrics = { version = "0.22", optional = true, default-features = false }
-metrics-exporter-prometheus = { version = "0.13", optional = true, default-features = false }
-ulid = "1.1.0"
-thiserror = "1"
-futures = "0.3"
-pin-project-lite = "0.2.13"
-# Needed only for sentry reporting
-uuid = { version = "1.6", optional = true }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = { version = "1", optional = true }
-
-[dependencies.tracing]
-default_features = false
-version = "0.1"
