--- packages/apalis-utils/src/codec/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "json")]
-pub mod json;
